+++ src/store/index.ts
@@ -13,7 +13,7 @@ const persistedReducers = persistCombineReducers<ApplicationState>(
     key: 'YaTA:store',
     migrate: createMigrate(migrations as any, { debug: false }),
     storage,
-    version: 7,
+    version: 8,
     whitelist: ['settings', 'user'],
   },
   reducers