+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Pause auto-scrolling temporarily when opening a context menu.
 - If enabled, messages from the user's chat history can be copied on double click.
 - If enabled, whispers can be copied on double click.
+- Add a new setting (enabled by default) to automatically focus the input field when focusing the application.
 
 ### 🐛 Bug Fix